requirements.txt

reflex==0.8.15a1
reflex-enterprise
psycopg2-binary